Rough draft on image gen

+++ projects/5-llm-image-gen/aws/deploy.sh
@@ -6,10 +6,21 @@ source $script_dir/aws.env
 
 project_dir_path="$server_home_dir/$project_dir_name"
 
-echo
-echo "Deploying code"
-echo
-rsync -a -P --include="docker-compose.yml" --include=".env" --include="app/***" --exclude="*" $script_dir/../ $ssh_connection:$project_dir_path
+# dry_run=true
+dry_run=false
+
+if [ "$dry_run" = true ]; then
+  echo
+  echo "Checking code diff"
+  echo
+  rsync -avn --itemize-changes -P --include="docker-compose.yml" --include=".env" --include="app/***" --exclude="*" $script_dir/../ $ssh_connection:$project_dir_path
+else
+  echo
+  echo "Deploying code"
+  echo
+  rsync -a -P --include="docker-compose.yml" --include=".env" --include="app/***" --exclude="*" $script_dir/../ $ssh_connection:$project_dir_path
+fi
+
 
 # echo
 # echo "Starting Docker Stack"
